options cell update, cleanup
